Move generic utility functions
